requirement.txt
moviepy==1.0.3
numpy==1.25.2
Pillow==10.0.0
requests==2.31.0
uvicorn==0.23.2
fastapi==0.100.1